+++ src/components/defaultComponents.js
@@ -2,7 +2,7 @@
 
 import Container from './Container';
 import Footer from './Footer';
-import Header from './Header';
+import Header, { HeaderButton } from './Header';
 import { Navigation, NavigationItem } from './Navigation';
 import View from './View';
 
@@ -10,6 +10,7 @@ export type CarouselComponents = {
   Container: typeof Container,
   Footer: typeof Footer,
   Header: typeof Header,
+  HeaderButton: typeof HeaderButton,
   Navigation: typeof Navigation,
   NavigationItem: typeof NavigationItem,
   View: typeof View,
@@ -19,6 +20,7 @@ export const components: CarouselComponents = {
   Container: Container,
   Footer: Footer,
   Header: Header,
+  HeaderButton: HeaderButton,
   Navigation: Navigation,
   NavigationItem: NavigationItem,
   View: View,
