+++ src/components/helpers.js
@@ -1,0 +1,4 @@
+// @flow
+
+export const smallDevice = '@media (max-width: 769px)';
+export const largeDevice = '@media (min-width: 770px)';
